Add initial open source release of libsalamander
Salamander is a secure messaging protocol for mobile devices.  It uses
the Axolotl Ratchet designed by Trevor Perrin and Moxie Marlinspike.

This library, libsalamander, is an implementation of the Salamander
protocol, and is used in the Silent Circle suite of applications.

Silent Circle is releasing libsalamander under the Apache License
Version 2.0 to help build a community around secure mobile messaging
and to encourage community review of the protocol and this
implementation.

This implementation is by Werner Dittmann.

The design of Salamander is a collaboration between Werner Dittmann,
Phil Zimmermann, and Travis Cross.

Thanks-to: Ronald Krueger <rkrueger@silentcircle.com>

...for packaging this library for public release.

Signed-off-by: Matt Neiderman <mneiderman@silentcircle.com>

...the General Counsel (CLO) for Silent Circle who authorized the
release of this library under the ASLv2.